Add true recursive identifier lookups
This commit also improves the semantics
of nil. To do so, comparisons of values
(e.g. `isEqual`) were moved out of the
`Eq` and `Ord` classes. Otherwise, the
semantics of comparing values that only
apply to evaluation would be applied to
compare the structure of two Haskell
values of type `Value` too. This in
undesirable, as it would make testing
difficult.